updates to style
