+++ src/components/key_gen/KeyGen.js
@@ -1,12 +1,11 @@
 import React, { useState } from "react";
 import { makeStyles } from "@material-ui/core/styles";
 import PassPhrase from "../shared/Passphrase";
-import Display from "../shared/BrowserResult";
 import HelpIcon from "@material-ui/icons/Help";
+import Result from "./Result";
 import {
   Box,
   Typography,
-  Button,
   TextField,
   Select,
   MenuItem,
@@ -38,24 +37,15 @@ const useStyles = makeStyles((theme) => ({
     },
   },
   formControl: { width: "100%" },
+
   main: {
-    paddingLeft: theme.spacing(2),
-    paddingRight: theme.spacing(2),
-    paddingBottom: theme.spacing(2),
+    marginLeft: theme.spacing(2),
+    padding: theme.spacing(2),
 
-    [theme.breakpoints.up(600 + theme.spacing(3) * 2)]: {
-      paddingLeft: theme.spacing(2),
-      paddingRight: theme.spacing(2),
-      paddingBottom: theme.spacing(2),
-    },
-  },
-
-  button: {
-    margin: "5px",
   },
 }));
 
-const KeyGen = (props) => {
+const KeyGen = () => {
   const classes = useStyles();
   const [key, setKey] = useState(undefined);
   const [bits, setBits] = useState(2048);
@@ -80,6 +70,7 @@ const KeyGen = (props) => {
   };
 
   let handleEmailValid = (email) => {
+    // eslint-disable-next-line
     let re = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
     if (re.test(email)) {
       return false;
@@ -129,12 +120,12 @@ const KeyGen = (props) => {
   };
 
   return (
-    <Box p={2}>
+    <Box>
       {typeof key == "undefined" ? (
         <>
           <form onSubmit={(e) => handleFormSubmit(e)}>
             <div className={classes.main}>
-              <Typography className={classes.heading} variant="h5" gutterBottom>
+              <Typography variant="h5" gutterBottom>
                 <b> RSA Key Generation</b>
               </Typography>
               <Box pb={4} pt={2}>
@@ -217,16 +208,18 @@ const KeyGen = (props) => {
                     </Grid>
                   </Grid>
                 </Box>
+                <Box mt={3}>
+                  <PassPhrase
+                    class={classes.pwInput}
+                    loading={loading}
+                    mainButtonText={"Generate"}
+                    modalButtonText={"Submit"}
+                    handleSubmit={handleFormSubmit}
+                    handleConfirm={handleConfirm}
+                  />
+                </Box>
               </div>
             </div>
-            <PassPhrase
-              class={classes.pwInput}
-              loading={loading}
-              mainButtonText={"Generate"}
-              modalButtonText={"Submit"}
-              handleSubmit={handleFormSubmit}
-              handleConfirm={handleConfirm}
-            />
           </form>
         </>
       ) : (
@@ -236,112 +229,4 @@ const KeyGen = (props) => {
   );
 };
 
-const Result = (props) => {
-  const classes = useStyles();
-
-  const [openPub, setOpenPub] = useState(false);
-  const [openPriv, setOpenPriv] = useState(false);
-
-  const dlKey = (key, name) => {
-    const element = document.createElement("a");
-    const file = new Blob([key], { type: "text/plain" });
-    element.href = URL.createObjectURL(file);
-    element.download = name + ".txt";
-    element.click();
-    element.remove();
-  };
-
-  let encKeys = props.encKeys;
-
-  return (
-    <div className={classes.main}>
-      <Box>
-        <Typography className={classes.heading} variant="h5" gutterBottom>
-          <b> Key Results </b>
-        </Typography>
-        <br />
-        <br />
-
-        <Typography
-          className={classes.heading}
-          color="primary"
-          variant="h6"
-          gutterBottom
-        >
-          Private Key
-        </Typography>
-
-        <Box pb={2}>
-          <Typography color="error">
-            <b>The private key is private. NEVER SHARE YOUR PRIVATE KEY</b>
-          </Typography>
-          <p>
-            Use the private key for decrypting data encrypted with your public
-            key.
-          </p>
-        </Box>
-        <Box pb={4}>
-          <Button
-            onClick={() => setOpenPriv(!openPriv)}
-            variant="outlined"
-            color={"primary"}
-            data-testid="privateKey"
-            className={classes.button}
-          >
-            {openPriv ? "Hide" : "In Browser"}
-          </Button>
-          <Button
-            onClick={() => dlKey(encKeys.privateKeyArmored, "private_key")}
-            variant="outlined"
-            color={"primary"}
-            className={classes.button}
-          >
-            Download
-          </Button>
-
-          {openPriv && (
-            <Display val={encKeys.privateKeyArmored} id="privateKey" />
-          )}
-        </Box>
-        <Typography
-          color="secondary"
-          className={classes.heading}
-          variant="h6"
-          gutterBottom
-        >
-          Public Key
-        </Typography>
-
-        <Box pb={2}>
-          <p>The public key is public. You could share it with anyone.</p>
-          <p>
-            The public key encrypts files that only your private key can
-            decrypt.
-          </p>
-        </Box>
-        <Box mb={2}>
-          <Button
-            onClick={() => setOpenPub(!openPub)}
-            variant="outlined"
-            color={"secondary"}
-            data-testid="publicKey"
-            className={classes.button}
-          >
-            {openPub ? "Hide" : "In Browser"}
-          </Button>
-          <Button
-            onClick={() => dlKey(encKeys.publicKeyArmored, "public_key")}
-            variant="outlined"
-            color={"secondary"}
-            className={classes.button}
-          >
-            Download
-          </Button>
-
-          {openPub && <Display val={encKeys.publicKeyArmored} id="publicKey" />}
-        </Box>
-      </Box>
-    </div>
-  );
-};
 export default KeyGen;
