+++ src/components/shared/KeyInput.js
@@ -30,6 +30,7 @@ const useStyles = makeStyles((theme) => ({
 
   main: {
     backgroundColor: "#FAFAFA",
+    borderRadius:'4px',
     padding: theme.spacing(2),
     [theme.breakpoints.up(600 + theme.spacing(3) * 2)]: {
       padding: theme.spacing(2),
