+++ src/index.js
@@ -6,13 +6,13 @@ import { HashRouter as Router } from "react-router-dom";
 import * as serviceWorker from "./serviceWorker";
 import { createMuiTheme, ThemeProvider } from "@material-ui/core/styles";
 import deepPurple from "@material-ui/core/colors/deepPurple";
-import teal from "@material-ui/core/colors/teal";
+import green from "@material-ui/core/colors/green";
 require('typeface-roboto')
 
 const theme = createMuiTheme({
   palette: {
     primary: deepPurple,
-    secondary: teal,
+    secondary: green,
     background: {
       default: "#ffffff",
     },
